requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
amqp==5.1.1 ; python_version >= '3.6'
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.2.2
automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==22.1.0
boto3==1.21.26
botocore==1.24.26
bs4==0.0.1
celery[redis]==5.2.7
certifi==2021.10.8
cffi==1.15.0
cfgv==3.4.0 ; python_version >= '3.8'
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
distlib==0.3.4
django==4.0.2
django-extensions==3.2.3
django-ninja==0.17.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
faker==19.6.2 ; python_version >= '3.8'
filelock==3.6.0
flake8==6.1.0
gensim==4.3.2
gunicorn==20.0.4
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
identify==2.5.29 ; python_version >= '3.8'
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
isort==5.10.1
jmespath==1.0.0
kombu==5.3.2 ; python_version >= '3.8'
mccabe==0.7.0 ; python_version >= '3.6'
msgpack==1.0.3
mypy-extensions==1.0.0 ; python_version >= '3.5'
mysqlclient==2.1.0
nodeenv==1.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.24.4 ; python_version >= '3.8'
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==3.4.0
prompt-toolkit==3.0.39 ; python_full_version >= '3.7.0'
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.0 ; python_version >= '3.8'
pycparser==2.21
pydantic==1.9.0
pyflakes==3.1.0 ; python_version >= '3.8'
pyjwt==2.8.0 ; python_version >= '3.7'
pymysql==1.0.2
pyopenssl==22.0.0
pyparsing==3.0.7
pysocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
redis==5.0.1
requests==2.27.1
s3transfer==0.5.2
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
selenium==4.1.3
service-identity==21.1.0
setuptools==68.2.2 ; python_version >= '3.8'
six==1.16.0
smart-open==6.4.0 ; python_version >= '3.6' and python_version < '4.0'
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
supervisor==4.2.4
tomli==2.0.1 ; python_version >= '3.7'
trio==0.20.0
trio-websocket==0.9.2
twisted==22.2.0
twisted-iocpsupport==1.0.4 ; platform_system == 'Windows'
txaio==22.2.1
types-pymysql==1.0.6
types-pytz==2021.3.6
types-pyyaml==6.0.5
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0 ; python_version >= '3.6'
virtualenv==20.14.0
wcwidth==0.2.8
wsproto==1.1.0
zope.interface==5.4.0
